requirements.txt
Flask
flask-cors
google-api-python-client
python-docx
google-auth
google-generativeai
gunicorn
